separated the logic for the favorite button and made it a single button, still some bugs
